refactor(Content): extract RssUtils
