+++ lib/feed_me/content.ex
@@ -9,8 +9,8 @@ defmodule FeedMe.Content do
   alias FeedMe.Content.Feed
   alias FeedMe.Content.FeedItem
   alias FeedMe.Repo
+  alias FeedMe.RssUtils
   alias FeedMe.YouTubeUtils
-  alias HTTPoison.Response
 
   @doc """
   Returns a list of feeds given a list of feed IDs.
@@ -235,55 +235,10 @@ defmodule FeedMe.Content do
     FeedItem.changeset(feed_item, attrs)
   end
 
-  def get_feed_from_rss_url(url_input) do
-    url =
-      if YouTubeUtils.is_youtube_channel_url(url_input) do
-        YouTubeUtils.get_rss_url_from_youtube_url(url_input)
-      else
-        url_input
-      end
-
-    %Response{body: body} = HTTPoison.get!(url)
-
-    case XmlToMap.naive_map(body) do
-      %{
-        "rss" => %{
-          "#content" => %{
-            "channel" => %{
-              "description" => description,
-              # "item" => items,
-              # "link" => link,
-              "title" => name
-            }
-          }
-        }
-      } ->
-        %{
-          name: name,
-          url: url,
-          description: description
-        }
-
-      %{
-        "feed" => %{
-          "title" => name,
-          "author" => %{
-            "uri" => description
-          }
-        }
-      } ->
-        %{
-          name: name,
-          url: url,
-          description: description
-        }
-    end
-  end
-
   def insert_all_feed_items(feed) do
     db_feed_items =
-      get_feed_items_from_rss_url(feed.url)
-      |> convert_rss_items_to_db_items(feed.id)
+      RssUtils.get_feed_items_from_rss_url(feed.url)
+      |> RssUtils.convert_rss_items_to_db_items(feed.id)
 
     Repo.insert_all(FeedItem, db_feed_items, on_conflict: :nothing)
   end
@@ -312,30 +267,6 @@ defmodule FeedMe.Content do
     |> Map.put(:description, :erlang.binary_to_term(item.description))
   end
 
-  defp get_feed_items_from_rss_url(url) do
-    %Response{body: body} = HTTPoison.get!(url)
-
-    case XmlToMap.naive_map(body) do
-      %{
-        "rss" => %{
-          "#content" => %{
-            "channel" => %{
-              "item" => items
-            }
-          }
-        }
-      } ->
-        items
-
-      %{
-        "feed" => %{
-          "entry" => items
-        }
-      } ->
-        items
-    end
-  end
-
   defp is_feed_item_read(item, user) do
     case AccountContent.get_feed_item_status(item.id, user.id) do
       [] ->
@@ -344,44 +275,6 @@ defmodule FeedMe.Content do
 
       [status = %AccountContent.FeedItemStatus{}] ->
         status.is_read
-    end
-  end
-
-  defp convert_rss_items_to_db_items(items, feed_id) do
-    Enum.map(items, fn item -> convert_rss_item_to_db_item(item, feed_id) end)
-  end
-
-  defp convert_rss_item_to_db_item(item, feed_id) do
-    case item do
-      %{
-        "title" => title,
-        "description" => description,
-        "link" => url,
-        "pubDate" => pub_date
-      } ->
-        %{
-          title: title,
-          description: :erlang.term_to_binary(description, [:compressed]),
-          url: url,
-          pub_date: pub_date,
-          feed_id: feed_id
-        }
-
-      %{
-        "title" => title,
-        "published" => pub_date,
-        "link" => %{
-          "#content" => description,
-          "-href" => url
-        }
-      } ->
-        %{
-          title: title,
-          description: :erlang.term_to_binary(description, [:compressed]),
-          url: url,
-          pub_date: pub_date,
-          feed_id: feed_id
-        }
     end
   end
 end
